Create Preprocessing text 2
Preprocessing text
Building a recommendation system, or any model, requires text to be preprocessed first.

A block of text from Sherlock Holmes is loaded here. Preprocess this text using the various techniques presented in the video to prepare it for further analysis.

The text variable is an excerpt from The Hound of the Baskervilles by Arthur Conan Doyle.

The following packages and functions have been loaded for you: nltk, torch, get_tokenizer, PorterStemmer, stopwords.

Instructions 2/3

Create a set of English stopwords and use list comprehension to filter these stop_words out of the text, making sure to ignore capitalization.