Create Embedding in PyTorch
Embedding in PyTorch
PyBooks found success with a book recommendation system. However, it doesn't account for some of the semantics found in the text. PyTorch's built-in embedding layer can learn and represent the relationship between words directly from data. Your team is curious to explore this capability to improve the book recommendation system. Can you help implement it?

torch and torch.nn as nn have been imported for you.

Instructions

Map a unique index to each word in words, saving to word_to_idx.
Convert word_to_idx into a PyTorch tensor and save to inputs.
Initialize an embedding layer using the torch module with ten dimensions.
Pass the inputs tensor to the embedding layer and review the output.