Create Building an LSTM model for text
Building an LSTM model for text
At PyBooks, the team is constantly seeking to enhance the user experience by leveraging the latest advancements in technology. In line with this vision, they have assigned you a critical task. The team wants you to explore the potential of another powerful tool: LSTM, known for capturing more complexities in data patterns. You are working with the same Newsgroup dataset, with the objective remaining unchanged: to classify news articles into three distinct categories:

rec.autos, sci.med, and comp.graphics.

The following packages have been loaded for you: torch, nn, optim.

Instructions

Set up an LSTM model by completing the LSTM and linear layers with the necessary parameters.
Initialize the model with the necessary parameters.
Train the LSTM model resetting the gradients to zero and passing the input data X_train_seq through the model.
Calculate the loss based on the predicted outputs and the true labels.