Create Evaluating RNN classification models
Evaluating RNN classification models
The team at PyBooks now wants you to evaluate the RNN model you created and ran using the Newsgroup dataset. Recall, the goal was to classify the articles into one of three categories:

rec.autos, sci.med, and comp.graphics.

The model was trained and you printed epoch and loss for each model.

Use torchmetrics to evaluate various metrics for your model. The following has been loaded for : Accuracy, Precision, Recall, F1Score.

An instance of rnn_model trained in the previous exercise in preloaded for you, too.

Instructions

Create an instance of each metric for multi-class classification with num_classes equal to the number of categories.
Generate the predictions for the rnn_model using the test data X_test_seq.
Calculate the metrics using the predicted classes and the true labels.